apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
redis==6.4.0
alembic==1.16.5
PyJWT==2.10.1
cryptography==45.0.7
PyNaCl==1.5.0
prometheus-client==0.23.1
